+++ .gitignore
@@ -65,3 +65,8 @@ $RECYCLE.BIN/
 /target/
 /target/
 /target/
+/target/
+/target/
+/bin/
+/bin/
+/target/
