+++ MbPets.iml
@@ -1,0 +1,46 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module org.jetbrains.idea.maven.project.MavenProjectsManager.isMavenModule="true" type="JAVA_MODULE" version="4">
+  <component name="NewModuleRootManager" LANGUAGE_LEVEL="JDK_1_7" inherit-compiler-output="false">
+    <output url="file://$MODULE_DIR$/target/classes" />
+    <output-test url="file://$MODULE_DIR$/target/test-classes" />
+    <content url="file://$MODULE_DIR$">
+      <sourceFolder url="file://$MODULE_DIR$/src/main/java" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/src/main/resources" type="java-resource" />
+      <excludeFolder url="file://$MODULE_DIR$/target" />
+    </content>
+    <orderEntry type="inheritedJdk" />
+    <orderEntry type="sourceFolder" forTests="false" />
+    <orderEntry type="library" name="Maven: org.spigotmc:spigot:1.8.3-R0.1-SNAPSHOT" level="project" />
+    <orderEntry type="library" name="Maven: net.zaiyers:UUIDDB:1.1-SNAPSHOT" level="project" />
+    <orderEntry type="library" name="Maven: org.mongodb:mongo-java-driver:2.11.4" level="project" />
+    <orderEntry type="library" name="Maven: net.md-5:bungeecord-api:1.8-SNAPSHOT" level="project" />
+    <orderEntry type="library" name="Maven: net.md-5:bungeecord-chat:1.8-SNAPSHOT" level="project" />
+    <orderEntry type="library" name="Maven: com.google.code.gson:gson:2.3.1" level="project" />
+    <orderEntry type="library" name="Maven: net.md-5:bungeecord-event:1.8-SNAPSHOT" level="project" />
+    <orderEntry type="library" name="Maven: net.md-5:bungeecord-protocol:1.8-SNAPSHOT" level="project" />
+    <orderEntry type="library" name="Maven: io.netty:netty-codec:4.0.24.Final" level="project" />
+    <orderEntry type="library" name="Maven: io.netty:netty-transport:4.0.24.Final" level="project" />
+    <orderEntry type="library" name="Maven: io.netty:netty-buffer:4.0.24.Final" level="project" />
+    <orderEntry type="library" name="Maven: io.netty:netty-common:4.0.24.Final" level="project" />
+    <orderEntry type="library" name="Maven: net.sf.trove4j:trove4j:3.0.3" level="project" />
+    <orderEntry type="library" name="Maven: net.md-5:bungeecord-config:1.8-SNAPSHOT" level="project" />
+    <orderEntry type="library" name="Maven: org.yaml:snakeyaml:1.14" level="project" />
+    <orderEntry type="library" name="Maven: org.bukkit:bukkit:1.8-R0.1-SNAPSHOT" level="project" />
+    <orderEntry type="library" name="Maven: commons-lang:commons-lang:2.6" level="project" />
+    <orderEntry type="library" name="Maven: com.googlecode.json-simple:json-simple:1.1.1" level="project" />
+    <orderEntry type="library" name="Maven: junit:junit:4.10" level="project" />
+    <orderEntry type="library" name="Maven: org.hamcrest:hamcrest-core:1.1" level="project" />
+    <orderEntry type="library" name="Maven: org.avaje:ebean:2.8.1" level="project" />
+    <orderEntry type="library" name="Maven: javax.persistence:persistence-api:1.0" level="project" />
+    <orderEntry type="library" name="Maven: com.comphenix.protocol:ProtocolLib:3.6.3-SNAPSHOT" level="project" />
+    <orderEntry type="library" name="Maven: cglib:cglib-nodep:2.2.2" level="project" />
+    <orderEntry type="library" name="Maven: com.comphenix.executors:BukkitExecutors:1.0.0" level="project" />
+    <orderEntry type="library" name="Maven: com.google.guava:guava:17.0" level="project" />
+    <orderEntry type="library" name="Maven: io.netty:netty-all:4.0.26.Final" level="project" />
+    <orderEntry type="library" name="Maven: LibsDisguises:LibsDisguises:8.3.1" level="project" />
+    <orderEntry type="library" name="Maven: net.milkbowl.vault:Vault:1.4.1" level="project" />
+    <orderEntry type="library" name="Maven: MbAnimalProtect:MbAnimalProtect:1.0" level="project" />
+    <orderEntry type="library" name="Maven: net.zaiyers.bukkit:UUIDDB:1.0" level="project" />
+    <orderEntry type="library" name="Maven: org.bukkit:craftbukkit:1.7.2-R0.1-SNAPSHOT" level="project" />
+  </component>
+</module>
